+++ src/mruby_integration/textures.cpp
@@ -25,6 +25,15 @@ mrb_value mrb_load_image(mrb_state *mrb, mrb_value) {
   return obj;
 }
 
+mrb_value mrb_unload_image(mrb_state *mrb, mrb_value) {
+  Image *image;
+  mrb_get_args(mrb, "d", &image, &Image_type);
+
+  UnloadImage(*image);
+
+  return mrb_nil_value();
+}
+
 mrb_value mrb_load_texture(mrb_state *mrb, mrb_value) {
   char *path;
   mrb_get_args(mrb, "z", &path);
@@ -108,6 +117,7 @@ mrb_value mrb_get_screen_data(mrb_state *mrb, mrb_value) {
 
 void append_textures(mrb_state *mrb) {
   mrb_define_method(mrb, mrb->kernel_module, "load_image", mrb_load_image, MRB_ARGS_REQ(1));
+  mrb_define_method(mrb, mrb->kernel_module, "unload_image", mrb_unload_image, MRB_ARGS_REQ(1));
 
   mrb_define_method(mrb, mrb->kernel_module, "load_texture", mrb_load_texture, MRB_ARGS_REQ(1));
   mrb_define_method(mrb, mrb->kernel_module, "unload_texture", mrb_unload_texture, MRB_ARGS_REQ(1));